+++ webapp/app/assessment/components/section/components/descriptions/components/analysisDescriptions.js
@@ -3,38 +3,30 @@ import PropTypes from 'prop-types'
 
 import { isPrintingMode, isPrintingOnlyTables } from '@webapp/app/assessment/components/print/printAssessment'
 
-import CommentableDescription from '@webapp/app/components/description/commentableDescription'
+import CommentableDescription from '@webapp/app/assessment/components/section/components/descriptions/components/commentableDescription'
 import useI18n from '@webapp/components/hooks/useI18n'
 
-const NationalDataDescriptions = props => {
+const AnalysisDescriptions = props => {
   const { section, disabled } = props
   const i18n = useI18n()
 
   return (
     !isPrintingOnlyTables() && (
       <div className="fra-description__container">
-        <h2 className="headline fra-description__group-header">{i18n.t('description.nationalData')}</h2>
+        <h2 className="headline fra-description__group-header">{i18n.t('description.analysisAndProcessing')}</h2>
         <CommentableDescription
-          title={i18n.t('description.dataSourcesPlus')}
+          title={i18n.t('description.estimationAndForecasting')}
           disabled={disabled}
           section={section}
-          name="dataSources"
+          name="estimationAndForecasting"
           showAlertEmptyContent={!isPrintingMode()}
           showDashEmptyContent={isPrintingMode()}
         />
         <CommentableDescription
-          title={i18n.t('description.nationalClassificationAndDefinitions')}
+          title={i18n.t('description.reclassification')}
           disabled={disabled}
           section={section}
-          name="nationalClassificationAndDefinitions"
-          showAlertEmptyContent={!isPrintingMode()}
-          showDashEmptyContent={isPrintingMode()}
-        />
-        <CommentableDescription
-          title={i18n.t('description.originalData')}
-          disabled={disabled}
-          section={section}
-          name="originalData"
+          name="reclassification"
           showAlertEmptyContent={!isPrintingMode()}
           showDashEmptyContent={isPrintingMode()}
         />
@@ -43,9 +35,9 @@ const NationalDataDescriptions = props => {
   )
 }
 
-NationalDataDescriptions.propTypes = {
+AnalysisDescriptions.propTypes = {
   section: PropTypes.string.isRequired,
   disabled: PropTypes.bool.isRequired,
 }
 
-export default NationalDataDescriptions
+export default AnalysisDescriptions
