+++ webapp/app/assessment/components/section/components/descriptions/components/description.less
