+++ webapp/app/assessment/components/section/components/descriptions/components/useDescription.js
@@ -1,4 +1,5 @@
 import { useEffect } from 'react'
+import { useDispatch } from 'react-redux'
 import * as R from 'ramda'
 
 import { debounce } from '@webapp/utils/functionUtils'
@@ -7,11 +8,13 @@ import useGetRequest from '@webapp/components/hooks/useGetRequest'
 import usePostRequest from '@webapp/components/hooks/usePostRequest'
 import useCountryIso from '@webapp/components/hooks/useCountryIso'
 import useOnUpdate from '@webapp/components/hooks/useOnUpdate'
+import * as autosave from '@webapp/app/components/autosave/actions'
 
 const getUrl = (countryIso, section, name) => `/api/country/descriptions/${countryIso}/${section}/${name}`
 const getData = name => R.pathOr(null, [name, 'content'])
 
 export default (name, section, template) => {
+  const dispatch = useDispatch()
   const countryIso = useCountryIso()
   const { data = template || null, setState, loading, dispatch: fetchData } = useGetRequest(
     getUrl(countryIso, section, name)
@@ -24,7 +27,9 @@ export default (name, section, template) => {
   })
 
   useOnUpdate(() => {
+    dispatch(autosave.start)
     debounce(postData, 'postDescriptionData', 800)()
+    dispatch(autosave.complete)
   }, [data])
 
   const onChange = content => {
