+++ webapp/app/assessment/components/traditionalTable/singleTraditionalTableView.js
@@ -7,9 +7,9 @@
 import React, { useEffect } from 'react'
 import { connect } from 'react-redux'
 import TraditionalTable from '@webapp/app/assessment/components/traditionalTable/traditionalTable'
-import NationalDataDescriptions from '@webapp/app/components/description/nationalDataDescriptions'
-import AnalysisDescriptions from '@webapp/app/components/description/analysisDescriptions'
-import GeneralComments from '@webapp/app/components/description/generalComments'
+import NationalDataDescriptions from '@webapp/app/assessment/components/section/components/descriptions/components/nationalDataDescriptions'
+import AnalysisDescriptions from '@webapp/app/assessment/components/section/components/descriptions/components/analysisDescriptions'
+import GeneralComments from '@webapp/app/assessment/components/section/components/descriptions/components/generalComments'
 import { fetchLastSectionUpdateTimestamp } from '@webapp/app/components/audit/actions'
 import DefinitionLink from '@webapp/components/definitionLink'
 import * as R from 'ramda'
