+++ webapp/app/assessment/fra/sections/carbonStock/carbonStockView.js
@@ -4,9 +4,9 @@ import * as R from 'ramda'
 
 import TraditionalTable from '@webapp/app/assessment/components/traditionalTable/traditionalTable'
 import tableSpec from './tableSpec'
-import NationalDataDescriptions from '@webapp/app/components/description/nationalDataDescriptions'
-import AnalysisDescriptions from '@webapp/app/components/description/analysisDescriptions'
-import GeneralComments from '@webapp/app/components/description/generalComments'
+import NationalDataDescriptions from '@webapp/app/assessment/components/section/components/descriptions/components/nationalDataDescriptions'
+import AnalysisDescriptions from '@webapp/app/assessment/components/section/components/descriptions/components/analysisDescriptions'
+import GeneralComments from '@webapp/app/assessment/components/section/components/descriptions/components/generalComments'
 import DefinitionLink from '@webapp/components/definitionLink'
 import ExcelCalculatorDownload from '@webapp/app/assessment/fra/sections/biomassStock/excelCalculatorDownload'
 import { fetchLastSectionUpdateTimestamp } from '@webapp/app/components/audit/actions'
