+++ webapp/app/assessment/fra/sections/extentOfForest/extentOfForestView.js
@@ -7,9 +7,9 @@ import Icon from '@webapp/components/icon'
 import DefinitionLink from '@webapp/components/definitionLink'
 import ChartWrapper from '@webapp/app/assessment/components/dataTable/chart/chartWrapper'
 import TableWithOdp  from '@webapp/app/assessment/fra/components/tableWithOdp'
-import NationalDataDescriptions from '@webapp/app/components/description/nationalDataDescriptions'
-import AnalysisDescriptions from '@webapp/app/components/description/analysisDescriptions'
-import GeneralComments from '@webapp/app/components/description/generalComments'
+import NationalDataDescriptions from '@webapp/app/assessment/components/section/components/descriptions/components/nationalDataDescriptions'
+import AnalysisDescriptions from '@webapp/app/assessment/components/section/components/descriptions/components/analysisDescriptions'
+import GeneralComments from '@webapp/app/assessment/components/section/components/descriptions/components/generalComments'
 import TraditionalTable from '@webapp/app/assessment/components/traditionalTable/traditionalTable'
 import NationalDataPointsPrintView
   from '@webapp/app/assessment/fra/sections/originalDataPoint/nationalDataPointsPrintView'
