--- webapp/app/components/description/index.js
@@ -1,1 +1,0 @@
-export { default } from './description'
